Safari-611.1.10/LayoutTests/js/dfg-putbyval-cfa-clobber-expected.txt

Tests that CFA knows that PutByVal can clobber the world.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS foo(array, -1) is 57
PASS successfullyParsed is true

TEST COMPLETE
